Slightly successful attempt to improve speed.

+++ digital_sky.py
@@ -32,10 +32,10 @@ class Datastore_Data(Resource):
         self.LAMP_LENGTH=LAMP_LENGTH # The length of each lamp module
         self.STRIP_LEDS=STRIP_LEDS   # The number of pixels at the start of each lamp which are special
         #FUTURE# self.strip_vals = full(int((self.LED_COUNT/self.LAMP_LENGTH)*self.STRIP_LEDS),4),0, dtype=uint8)
-        self.strip_vals = array([0,0,0,0], dtype=int)
+        self.strip_vals = array([0,0,0,0], dtype=uint8)
         self.master_brightness = 1.0
         self.rgbw_brightness = [1.0,1.0,1.0,1.0]
-        self.strips = full((self.LED_COUNT,4),0, dtype=uint8)
+        self.strips = full((self.LED_COUNT,4),0)
         self.plugins = []
         self.animations=[]
         self.strip_animations=[]
